# docs/swagger.yaml
basePath: /
definitions:
  controllers.AuthRequest:
    properties:
      email:
        example: haidar@gmail.com
        type: string
      password:
        example: haidarpassword
        type: string
    type: object
  controllers.AuthResponse:
    properties:
      token:
        example: xxxxxxxxxxxxxxx...
        type: string
    type: object
  controllers.ErrorResponse:
    properties:
      error:
        example: Invalid Email or Password
        type: string
    type: object
  controllers.ExerciseDelReq:
    properties:
      id:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - id
    type: object
  controllers.ExerciseRequest:
    properties:
      category:
        type: string
      equipment:
        type: string
      muscleGroup:
        type: string
      name:
        type: string
    type: object
  controllers.ExerciseUpdateRequest:
    properties:
      category:
        type: string
      equipment:
        type: string
      id:
        type: integer
      muscleGroup:
        type: string
      name:
        type: string
    required:
    - id
    type: object
  controllers.GoogleTokenRequest:
    properties:
      id_token:
        type: string
    required:
    - id_token
    type: object
  controllers.SignupRequest:
    properties:
      email:
        example: haidar@gmail.com
        type: string
      fullName:
        example: Haidar Sebastian
        maxLength: 24
        minLength: 3
        type: string
      password:
        example: haidarpassword
        maxLength: 24
        minLength: 8
        type: string
      username:
        example: haidarIron
        maxLength: 24
        minLength: 3
        type: string
    required:
    - email
    - fullName
    - password
    - username
    type: object
  controllers.updateReq:
    properties:
      fullName:
        type: string
      id:
        minimum: 1
        type: integer
      username:
        type: string
    required:
    - id
    type: object
host: api-staging.liftlogs.my.id
info:
  contact: {}
  description: Backend API for IronLog
  title: IronLog Backend Go API
  version: "1.0"
paths:
  /api/v1/auth/signin:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Login a user
      tags:
      - auth
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Signup credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/v1/exercises:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Exercise data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ExerciseDelReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Delete exercises
      tags:
      - exercises
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all exercises
      tags:
      - exercises
    patch:
      consumes:
      - application/json
      parameters:
      - description: Exercise data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ExerciseUpdateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Update a exercise
      tags:
      - exercises
    post:
      consumes:
      - application/json
      parameters:
      - description: Exercise data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ExerciseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create a new exercise
      tags:
      - exercises
  /api/v1/exercises/{id}:
    get:
      parameters:
      - description: Exercise ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a single exercise
      tags:
      - exercises
  /api/v1/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all users
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: delete a user
      tags:
      - users
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a user
      tags:
      - users
    patch:
      parameters:
      - description: user data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.updateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a user
      tags:
      - users
  /auth/google/mobile:
    post:
      consumes:
      - application/json
      description: Authenticate user via Google ID token (mobile flow) and return
        a JWT token
      parameters:
      - description: Google ID Token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.GoogleTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns JWT token and user info
          schema:
            additionalProperties: true
            type: object
        "400":
          description: id_token is required
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid Google token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Sign in with Google
      tags:
      - auth
  /healthCheck:
    get:
      description: Perform status check to the server
      responses: {}
      summary: Check server status
schemes:
- https
swagger: "2.0"